+++ cli/tests/pcluster/api/controllers/test_cluster_operations_controller.py
@@ -68,6 +68,9 @@ HeadNode:
   InstanceType: t2.micro
   Networking:
     SubnetId: subnet-12345678
+  LocalStorage:
+    RootVolume:
+      Size: 80
   Ssh:
     KeyName: ec2-key-name
 Scheduling: