+++ cli/tests/pcluster/config/test_cluster_config.py
@@ -19,6 +19,7 @@ from pcluster.config.cluster_config import (
     HeadNodeSsh,
     HealthChecks,
     Image,
+    LocalStorage,
     LoginNodes,
     LoginNodesImage,
     LoginNodesNetworking,
@@ -26,6 +27,7 @@ from pcluster.config.cluster_config import (
     LoginNodesSsh,
     PlacementGroup,
     QueueImage,
+    RootVolume,
     SharedEbs,
     SlurmClusterConfig,
     SlurmComputeResource,
@@ -195,7 +197,9 @@ class TestBaseClusterConfig:
         return BaseClusterConfig(
             cluster_name="clustername",
             image=Image("alinux2"),
-            head_node=HeadNode("c5.xlarge", HeadNodeNetworking("subnet")),
+            head_node=HeadNode(
+                "c5.xlarge", HeadNodeNetworking("subnet"), local_storage=LocalStorage(root_volume=RootVolume(size=50))
+            ),
         )
 
     @pytest.fixture()
@@ -203,7 +207,9 @@ class TestBaseClusterConfig:
         return SlurmClusterConfig(
             cluster_name="clustername",
             image=Image("alinux2"),
-            head_node=HeadNode("c5.xlarge", HeadNodeNetworking("subnet")),
+            head_node=HeadNode(
+                "c5.xlarge", HeadNodeNetworking("subnet"), local_storage=LocalStorage(root_volume=RootVolume(size=50))
+            ),
             scheduling=SlurmScheduling(
                 [
                     SlurmQueue(