+++ CHANGELOG.md
@@ -1,6 +1,17 @@
 CHANGELOG
 =========
 
+3.9.0
+------
+
+**ENHANCEMENTS**
+- Add support for installing Intel OneAPI Base Toolkit and HPC Toolkit, and Intel Python.
+  - Intel OneAPI Base Toolkits: 2023.2.0
+  - Intel OneAPI HPC Toolkits: 2023.2.0
+  - Intel Python: 2023.2.0
+  - Critical Update for Intel oneAPI DPC++/C++ Compiler: 2023.2.1
+  - Critical Update for Intel Fortran Compiler & Intel® Fortran Compiler Classic: 2023.2.1
+
 3.8.0
 ------